app.py
import streamlit as st
from aip import AipOcr
from openpyxl import load_workbook
from openpyxl.drawing.image import Image as XLImage
from PIL import Image as PILImage
import io

# --- 1. 网页配置 ---
st.set_page_config(page_title="超市价签自动识别系统")
st.title("🥬 蔬菜价签自动识别")

# 在侧边栏配置密钥，更安全
with st.sidebar:
    st.header("API 配置")
    app_id = st.text_input("APP_ID", type="a7e8877dba554481becdbf135ce671ee")
    api_key = st.text_input("API_KEY", type="f59OoSeqYPqMUD1tecX1yeC0")
    secret_key = st.text_input("SECRET_KEY", type="VF87Je72q32LDe3q75VH9RI7CQbzIb4G")

# --- 2. 核心逻辑 (保持你之前的 V4 逻辑) ---
def process_ocr(img_bytes, client):
    res = client.accurate(img_bytes)
    # ... 这里插入你之前的识别逻辑 (get_ocr_info) ...
    # 记得返回品名和价格
    return "品名", "价格"

# --- 3. 网页上传界面 ---
uploaded_template = st.file_uploader("第一步：上传 Excel 模板", type=['xlsx'])
uploaded_imgs = st.file_uploader("第二步：上传价签照片", type=['jpg', 'png', 'jpeg'], accept_multiple_files=True)

if st.button("开始识别并生成报表"):
    if not (uploaded_template and uploaded_imgs and app_id):
        st.error("请确保填写了 API 密钥并上传了所有文件！")
    else:
        client = AipOcr(app_id, api_key, secret_key)
        wb = load_workbook(io.BytesIO(uploaded_template.read()))
        ws = wb.active
        
        # 处理逻辑...
        st.success("处理完成！")
        
        # 提供下载按钮
        output = io.BytesIO()
        wb.save(output)
        st.download_button(label="📥 下载识别好的 Excel", data=output.getvalue(), file_name="result.xlsx")