requirements.txt
streamlit
baidu-aip
openpyxl
Pillow